added cells layout

+++ NUIKit/NUIKit/NUILayout/NUIGridLayoutItem.m
@@ -7,6 +7,7 @@
 //
 
 #import "NUIGridLayoutItem.h"
+
 #import "UIView+NUILayout.h"
 
 @implementation NUIGridLayoutItem
@@ -36,22 +37,22 @@
     self.view.needsToUpdateSize = YES;
 }
 
-- (void)setColumn:(int)column
+- (void)setColumn:(NSUInteger)column
 {
     self.columnRange = (NSRange){column, self.columnRange.length};
 }
 
-- (int)column
+- (NSUInteger)column
 {
     return columnRange_.location;
 }
 
-- (void)setRow:(int)row
+- (void)setRow:(NSUInteger)row
 {
     self.rowRange = (NSRange){row, self.rowRange.length};
 }
 
-- (int)row
+- (NSUInteger)row
 {
     return rowRange_.location;
 }
